+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - RP-Inititated Logout support
 
+### Changed
+
+- Rename `ClaimsResolveEvent` to `UserClaimsResolveEvent`
+
 ## [0.1.0] - 2025-07-07
 
 ### Added